+++ src/components/Header/element.js
@@ -1,6 +1,6 @@
 import { ID } from '../../constants/selector.js';
 import { MACHINE } from '../../constants/machine.js';
-import { Nav, ButtonWithId } from '../elements.js';
+import { Nav, ButtonWithId } from '../Element/elements.js';
 
 export const createTabNav = (event) => {
   const tabNav = Nav(ID.NAV);
